pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ttvton"
dynamic = ["version"]
description = "TatVTON — Tattoo Virtual Try-On: realistic tattoo compositing on body photos using SAM + SDXL Inpainting + ControlNet + IP-Adapter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "TatVTON Contributors" }]
keywords = ["tattoo", "virtual-try-on", "diffusion", "inpainting", "SAM"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "Pillow>=9.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
densepose = [
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/khope/tatvton"
Repository = "https://github.com/khope/tatvton"
Issues = "https://github.com/khope/tatvton/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "no-guess-dev"
fallback_version = "0.0.0"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ttvton"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: marks tests that require a GPU (deselect with '-m \"not gpu\"')",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
